Updated Django Version, Attempt at Root URL Compliance
Updated to Django 1.2 and reworked data structure so that versions and
rankings are now pointing to packages (rather than a m2m from within the
package model).

Also, this is a first (and incomplete) attempt at complying with the CommonJS
root URL where multiple verions are presented. (This updates the HTML output
as well, but the big focus was on JSON.)